vpr: Move MapLookahead class into router_lookahead_map.h/.cpp
