+++ vpr/src/route/router_lookahead.h
@@ -68,14 +68,6 @@ class ClassicLookahead : public RouterLookahead {
     float classic_wire_lookahead_cost(int node, int target_node, float criticality, float R_upstream) const;
 };
 
-class MapLookahead : public RouterLookahead {
-  protected:
-    float get_expected_cost(int node, int target_node, const t_conn_cost_params& params, float R_upstream) const override;
-    void compute(const std::vector<t_segment_inf>& segment_inf) override;
-    void read(const std::string& file) override;
-    void write(const std::string& file) const override;
-};
-
 class NoOpLookahead : public RouterLookahead {
   protected:
     float get_expected_cost(int node, int target_node, const t_conn_cost_params& params, float R_upstream) const override;
